Transforming the project into a proper packagable layout
