+++ .gitignore
@@ -1,3 +1,4 @@
 venv
 *.pyc
 wadl.xml
+*.egg-info
