+++ instagram_scraper/__init__.py
@@ -1,0 +1,1 @@
+from instagram_scraper.app import InstagramScraper
