--- requirements.txt
@@ -1,3 +1,0 @@
-requests>=1.0.4
-futures>=2.2.0
-tqdm~=3.8.0
